Split into two files... struggling
for some reason the executeTime in home.JS does not work, and I'm not sure why... I've tried importing and exporting it multiple ways and it doesn't work at all.
