+++ home.html
@@ -47,7 +47,7 @@
                 <p id="date"></p>
             </div>
         </div>
-        <script src="home.js"></script>
+        <script type="module" src="home.js"></script>
 
         <div class="column">
             <h1 style="background-color: white;"><em>Time in Other Regions</em></h1>
