+++ home.js
@@ -1,3 +1,5 @@
+import * as DC from './dateCheck.js';
+
 function executeTime(){
     const hour = ReadHour();
     const minute = ReadMinute();
@@ -25,7 +27,8 @@ function executeTime(){
 
     document.getElementById("remTime").innerText = `${remHours} hours, ${remMinutes} minutes.`;
 
-    DateCheck();
+    DC.DateCheck();
+
 }
 
 
@@ -57,52 +60,5 @@ function ReadMeridiem(){
     return document.getElementById("meridiem").value;
 }
 
-function DateCheck() {
-    let month = ReadMonth();
-    let day = ReadDay();
-    const yearInput = document.getElementById("year").value;
 
-    const year = parseInt(yearInput, 10);
-
-    //days in each month Janauary to December
-    const daysInMonth = [31, 28, 31, 30, 31, 30, 31, 31, 30, 31, 30, 31];
-    let currentDays;
-
-    //need to set currentDays to the amount of days that are provided in the month
-    for(let i = 0; i < daysInMonth.length; i++){
-        if(month == i) {
-            currentDays = daysInMonth[i-1];
-        }
-    }
-
-    document.getElementById("date").innerText = `${month} ${currentDays}`;
-}
-
-//function to read the month
-function ReadMonth(){
-    const monthInput = document.getElementById("month").value;
-    //checking to see if month is valid
-    const month = parseInt(monthInput, 10);
-
-    if(month >= 0 && month <= 12){
-        return month;
-    }else {
-        alert("Invalid month. Type in a month from 1-12");
-        return null;
-    }
-}
-
-//function to read the days
-function ReadDay(){
-    const dayInput = document.getElementById("day").value;
-    //checking to see if day is valid
-    const day = parseInt(dayInput, 10);
-
-    if(day >= 0 && day <= 12){
-        return day;
-    }else {
-        alert("Invalid day. Please type in a day that is within the proper range for the month.");
-        return null;
-    }
-}
